+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Habit autocomplete definition refinements
 - Optional titles in habit rules
+- Habit autocompletion type other habit rule
+
 ### Changed:
 - New Flutter version & dependencies
 - Habits displayed for last 30 days by default on both mobile and desktop